+++ src/masterpass/internal/loading-frame.js
@@ -1,5 +1,6 @@
 'use strict';
 
+var sanitizeUrl = require('@braintree/sanitize-url').sanitizeUrl;
 var querystring = require('../../lib/querystring');
 
 function noop() {}
@@ -12,12 +13,13 @@ function start() {
   config.failureCallback = noop;
   config.cancelCallback = noop;
   config.successCallback = noop;
+  config.callbackUrl = sanitizeUrl(config.callbackUrl);
 
   script.type = 'text/javascript';
   script.src = 'https://' + envSubdomain + 'static.masterpass.com/dyn/js/switch/integration/MasterPass.client.js';
 
   script.onload = function () {
-    window.MasterPass.client.checkout(config);
+    global.MasterPass.client.checkout(config);
   };
 
   document.body.appendChild(script);